fix: new approach to setup user-env on UNIX
An 'env' file is created now, heavily inspired by the rust's cargo
approach (See #65), located in '~/.config/godotenv/env'. It's
generated by GodotEnv and contains commands in 'sh' syntax to setup
the user shell.

The new 'env' file is responsible for:
- Prepend the GodotEnv bin folder path into PATH;
- 'Export' GODOT env-var pointing to Godot symlink.

We expect all POSIX compatible shells to work with it.

Fixes #65