feat: version-free install and use commands
Enables issuing "godot install" and "godot use" without a version
parameter. If no version parameter is provided, the command will walk
up the directory tree in search of files named "global.json",
"*.csproj", and ".godotrc". A global.json at any directory level will be
used before a csproj, and a csproj at any level will be used before a
.godotrc. A file closer to the working directory will be preferred to
one of the same type further up the tree. The first file according to
this ordering scheme that contains a valid Godot version will be used to
infer the version for the command.

.godotrc files must contain a valid version string, either in GodotSharp
style or in Godot release style, as their first line.

Versions from global.json files or csproj files are assumed to specify
.NET versions of Godot. A .godotrc file may specify a non-.NET version
by preceding the version string with a tilde (~) character.

* Adds a GodotVersionSpecifierRepository to search files on disk for
  valid Godot version specifiers.
* Adds an IGodotVersionFile interface and types implementing it for
  global.json, csproj, and .godotrc files.
* Extends FileClient to support reading the various file types.
* Extends GodotContext to include the GodotVersionSpecifierRepository.
* Extends Main to construct the GodotVersionSpecifierRepository.
* Makes the RawVersion/Version parameter optional for
  GodotInstallCommand and GodotUseCommand, respectively, and extends
  each command to search for version-specifier files using the
  GodotVersionSpecifierRepository.