fix: use " no-dotnet" instead of "~" in .godotrc
Changes the representation of non-.NET versions in .godotrc files.
Instead of prefixing the version string with "~", we now append
" no-dotnet" (including space) to the version string to match the flag
used for the install and use commands.

(We also quietly accept " non-dotnet" and " not-dotnet", but
" no-dotnet" is the official specifier.)

Updates tests and README to reflect the change.

Also updates the output of the list command to use "no-dotnet" instead
of "not-dotnet" to designate installed non-.NET versions, to better
match the specifiers used elsewhere.