feat: add "godot pin" to pin project version
Adds a "godotenv godot pin" command which pins the currently active
Godot version as the preferred version for the project where the command
is executed. This command creates a version-specifying file in the top-
level project directory. This file can be used later by the version-free
"godot install" and "godot use" commands to use the pinned Godot
version.

To find the top-level project directory, the command walks up the
directory tree from the working directory and selects the highest
directory containing a ".sln" file, if one exists, or the highest
directory containing a "project.godot" file if no ".sln" file exists in
the directory tree.

If the currently active Godot version is .NET-enabled, the command
writes the version as a Godot.NET.Sdk value in a "global.json" file in
the top-level project directory. (If such a file exists, the value is
written into it; if not, a new file is created.)

If the currently active Godot version is non-.NET, the command writes
the version into a ".godotrc" file in the top-level project directory.
The contents of any existing file are erased.

- Adds a new method to IVersionSerializer to serialize versions
  including their .NET status, to standardize the behavior across
  repositories.
- Adds an abstract base class VersionSerializer to provide a default
  implementation of the new IVersionSerializer method without requiring
  casting to the interface. (The interface is retained for consistency
  with IVersionDeserializer.)
- Adds a method to IGodotVersionFile to write a Godot version into a
  version-specifying file, and implementations for all version-
  specifying file types. (CsprojFile's implementation throws a
  NotSupportedException.)
- Adds methods to FileClient to support writing Godot versions to file.
- Adds methods to GodotVersionSpecifierRepository to find the top-level
  project directory and to pin the current version in a version-
  specifying file in a given directory, using the new IGodotVersionFile
  functionality.
- Adds a new GodotPinCommand that uses the new
  GodotVersionSpecifierRepository functionality to pin the currently
  active Godot version in the project where the command is executed.